+++ docs/constants/navigation.js
@@ -162,7 +162,6 @@ const general = [
       'Bundling',
       [
         makePage('guides/customizing-metro.mdx'),
-        makePage('guides/react-compiler.mdx'),
         makePage('guides/analyzing-bundles.mdx'),
         makePage('guides/tree-shaking.mdx'),
         makePage('guides/minify.mdx'),
@@ -516,7 +515,7 @@ const learn = [
 const preview = [
   makeSection('Preview', [
     makePage('preview/introduction.mdx'),
-    makePage('preview/support.mdx'),
+    makePage('preview/react-compiler.mdx'),
     { expanded: true },
   ]),
 ];