+++ packages/expo/src/Expo.ts
@@ -16,3 +16,5 @@ export {
   requireOptionalNativeModule,
   reloadAppAsync,
 } from 'expo-modules-core';
+
+export { useEvent } from './hooks/useEvent';